change project structure to be feature based
