+++ src/App.tsx
@@ -3,19 +3,19 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import "./index.css";
 import Layout from "./components/Layout/Layout";
-import LoginPage from "./pages/Login/LoginPage";
+import LoginPage from "./features/auth/pages/LoginPage";
 import HomePage from "./pages/Home/HomePage";
-import RegisterPage from "./pages/Register/RegisterPage";
+import RegisterPage from "./features/auth/pages/RegisterPage";
 import StartPage from "./pages/Start/StartPage";
-import GameDetailPage from "./pages/GameDetail/GameDetailPage";
-import GameListPage from "./pages/GameList/GameListPage";
-import UserProfilePage from "./pages/UserProfile/UserProfilePage";
-import NotificationsPage from "./pages/Notifications/NotificationsPage";
-import UserSettingsPage from "./pages/UserSettings/UserSettingsPage";
+import GameDetailPage from "./features/games/pages/GameDetailPage";
+import GameListPage from "./features/games/pages/GameListPage";
+import UserProfilePage from "./features/users/pages/UserProfilePage";
+import NotificationsPage from "./features/notifications/pages/NotificationsPage";
+import UserSettingsPage from "./features/users/pages/UserSettingsPage";
 import NotFound from "./pages/NotFound/NotFound";
 import SearchEnginePage from "./pages/SearchEngine/SearchEnginePage";
-import RequireAuth from "./helpers/RequireAuth";
-import { AuthProvider } from "./context/AuthProvider";
+import RequireAuth from "./features/auth/components/RequireAuth";
+import { AuthProvider } from "./features/auth/context/AuthProvider";
 
 function App(): React.JSX.Element {
   return (
