--- src/context/AuthProvider.tsx
@@ -1,37 +1,0 @@
-import React, { createContext, useContext, useState, ReactNode } from "react";
-import { LocalStorageUserType } from "@/helpers/CustomTypes";
-
-interface AuthContextType {
-  user: LocalStorageUserType | null;
-  login: (user: LocalStorageUserType) => void;
-  logout: () => void;
-}
-
-const AuthContext = createContext<AuthContextType | undefined>(undefined);
-
-export function AuthProvider({ children }: { children: ReactNode }) {
-  const [user, setUser] = useState<LocalStorageUserType | null>(() => {
-    const savedUser = localStorage.getItem("user");
-    return savedUser ? JSON.parse(savedUser) : null;
-  });
-
-  const login = (userData: LocalStorageUserType) => {
-    localStorage.setItem("user", JSON.stringify(userData));
-    setUser(userData);
-  };
-
-  const logout = () => {
-    localStorage.removeItem("user");
-    setUser(null);
-  };
-
-  return <AuthContext.Provider value={{ user, login, logout }}>{children}</AuthContext.Provider>;
-}
-
-export function useAuth() {
-  const context = useContext(AuthContext);
-  if (context === undefined) {
-    throw new Error("useAuth must be used within an AuthProvider");
-  }
-  return context;
-}
